refactor: 구입 금액 관련 로직을 purchaseView로 분리
